+++ packages/insomnia/src/ui/routes/actions.tsx
@@ -19,7 +19,6 @@ import { isRequestGroup, isRequestGroupId } from '../../models/request-group';
 import { isRequestGroupMeta } from '../../models/request-group-meta';
 import type { UnitTest } from '../../models/unit-test';
 import type { UnitTestSuite } from '../../models/unit-test-suite';
-import type { WorkspaceMeta } from '../../models/workspace-meta';
 import { getSendRequestCallback } from '../../network/unit-test-feature';
 import { insomniaFetch } from '../../ui/insomniaFetch';
 import { invariant } from '../../utils/invariant';
@@ -39,37 +38,6 @@ export function safeToUseInsomniaFileName(fileName: string) {
 export function safeToUseInsomniaFileNameWithExt(fileName: string) {
   return `${safeToUseInsomniaFileName(fileName)}.yaml`;
 }
-
-export const moveWorkspaceIntoProjectAction: ActionFunction = async ({ request, params }) => {
-  const { organizationId } = params;
-
-  invariant(typeof organizationId === 'string', 'Organization ID is required');
-
-  const formData = await request.formData();
-  const projectId = formData.get('projectId');
-  const workspaceId = formData.get('workspaceId');
-  invariant(typeof projectId === 'string', 'Project ID is required');
-  const project = await models.project.getById(projectId);
-  invariant(project, 'Project not found');
-
-  invariant(typeof workspaceId === 'string', 'Workspace ID is required');
-  const workspace = await models.workspace.getById(workspaceId);
-  invariant(workspace, 'Workspace not found');
-
-  await models.workspace.update(workspace, {
-    parentId: projectId,
-  });
-
-  return null;
-};
-
-export const updateWorkspaceMetaAction: ActionFunction = async ({ request, params }) => {
-  const { workspaceId } = params;
-  invariant(typeof workspaceId === 'string', 'Workspace ID is required');
-  const patch = (await request.json()) as Partial<WorkspaceMeta>;
-  await models.workspaceMeta.updateByParentId(workspaceId, patch);
-  return null;
-};
 
 // Test Suite
 export const createNewTestSuiteAction: ActionFunction = async ({ request, params }) => {
